rename muxrpc packages
