--- muxrpc/doc.go
@@ -1,4 +1,0 @@
-// Package muxrpc contains the muxrpc handler packages for the room server.
-//
-// The implementation the actual multiplexing implementation is go.cryptoscope.co/muxrpc.
-package muxrpc
