+++ muxrpc/handlers/doc.go
@@ -1,0 +1,4 @@
+// Package handlers contains the muxrpc handler packages for the room server.
+//
+// The implementation the actual multiplexing implementation is go.cryptoscope.co/muxrpc.
+package handlers
