+++ muxrpc/test/nodejs/testscripts/simple_client.js
