+++ muxrpc/test/nodejs/testscripts/template.js
